rename server.js into app.js
